# AstraServerGo.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AstraServerGo",
    "description": "基于 `AstraScheduleServerGo` 源码生成的 OpenAPI 3.0 文档。\n\n- 共 55 个 API 端点，覆盖客户端课表、天气、WebSocket、管理后台认证、用户管理、学校/年级/班级管理、配置管理、自动任务、倒数日、备份导入导出等功能。\n- JWT 认证接口使用 `jwtAuth` 安全方案（通过 POST /web/auth/login 获取 Token）。\n- 写入接口同时支持 BasicAuth（兼容旧版客户端）和 JWT + 密码验证。\n- 部分接口返回结构是动态 JSON（`map[string]interface{}`），已尽量按实际行为建模。\n",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/web/backup/export": {
      "get": {
        "summary": "导出完整备份",
        "description": "导出数据库中所有配置数据（课表、作息、科目、设置、版本、自动任务、倒数日）为 JSON 文件下载。需要 BasicAuth 认证。",
        "operationId": "exportBackup",
        "tags": ["web-backup"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "JSON 备份文件下载（Content-Disposition: attachment）",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BackupPayload" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "导出失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/backup/import": {
      "post": {
        "summary": "导入备份",
        "description": "导入备份 JSON 文件。支持两种方式：\n1. multipart/form-data 上传 `file` 字段（.json 文件，最大 50MB）\n2. 直接发送 JSON 请求体\n导入模式为 overwrite（覆盖已存在的记录）。需要 BasicAuth 认证。",
        "operationId": "importBackup",
        "tags": ["web-backup"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BackupPayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "导入成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "备份导入完成" },
                    "data": { "$ref": "#/components/schemas/BackupImportResult" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "导入失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/backup/full-export": {
      "post": {
        "summary": "完整备份导出",
        "description": "与 `GET /web/backup/export` 功能相同，导出全量数据为 JSON 文件下载。使用 POST 方法以支持更灵活的参数扩展。需要 BasicAuth 认证。",
        "operationId": "fullExportBackup",
        "tags": ["web-backup"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "JSON 备份文件下载",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BackupPayload" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "导出失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/backup/full-import": {
      "post": {
        "summary": "完整备份导入（支持 overwrite/skip 模式）",
        "description": "与 `POST /web/backup/import` 类似，但支持通过 `mode` 参数控制导入策略：\n- `overwrite`（默认）：覆盖已存在的记录\n- `skip`：跳过已存在的记录\n支持 multipart/form-data 上传或直接 JSON 请求体。需要 BasicAuth 认证。",
        "operationId": "fullImportBackup",
        "tags": ["web-backup"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BackupPayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "导入成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "mode", "data"],
                  "properties": {
                    "status": { "type": "integer", "example": 200 },
                    "message": { "type": "string", "example": "备份导入完成" },
                    "mode": { "type": "string", "enum": ["overwrite", "skip"] },
                    "data": { "$ref": "#/components/schemas/BackupImportResult" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "Basic Auth 认证失败" },
          "500": {
            "description": "导入失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailyClass": {
        "description": "一天的课表（客户端格式，classList 为一维数组）",
        "type": "object",
        "properties": {
          "Chinese": { "type": "string", "description": "星期中文名" },
          "English": { "type": "string", "description": "星期英文名" },
          "classList": {
            "type": "array",
            "items": { "type": "string" },
            "description": "每节课的科目简称列表"
          },
          "timetable": { "type": "string", "description": "使用的作息表名称" }
        }
      },
...
      "TimetableMap": {
        "type": "object",
        "description": "作息表映射：键为作息名称（如「常日」），值为时间范围→课节索引/标签的映射。值可以是整数（课节号）、字符串（标签如「课间」）或数字。",
        "additionalProperties": {
          "type": "object",
          "additionalProperties": {
            "oneOf": [
              { "type": "integer" },
              { "type": "number" },
              { "type": "string" }
            ]
          }
        }
      },
...
      "CountdownScheduleItem": {
        "type": "object",
        "required": ["name", "date", "priority"],
        "properties": {
          "name": { "type": "string", "description": "倒数日名称" },
          "date": { "type": "string", "format": "date", "description": "目标日期（YYYY-MM-DD）" },
          "priority": { "type": "integer", "description": "优先级，数字越大优先级越高" }
        }
      },
...
      "BackupMeta": {
        "description": "备份元数据",
        "type": "object",
        "properties": {
          "schema_version": { "type": "integer", "description": "备份格式版本号" },
          "generated_at": { "type": "string", "description": "备份生成时间（RFC3339 格式）" }
        }
      },
      "BackupPayload": {
        "description": "完整备份数据结构，包含数据库中所有配置表的全量数据",
        "type": "object",
        "properties": {
          "meta": { "$ref": "#/components/schemas/BackupMeta" },
          "schedules": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackupScheduleRecord" },
            "description": "课表配置"
          },
          "client_configs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackupClientConfigRecord" },
            "description": "客户端通用设置"
          },
          "timetables": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackupTimetableRecord" },
            "description": "作息配置"
          },
          "subjects": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackupSubjectRecord" },
            "description": "科目配置"
          },
          "data_versions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackupDataVersionRecord" },
            "description": "数据版本"
          },
          "autorun_records": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackupAutorunRecord" },
            "description": "自动任务记录"
          },
          "countdown_records": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BackupCountdownRecord" },
            "description": "倒数日配置"
          }
        }
      },
      "BackupScheduleRecord": {
        "description": "数据库课表记录",
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "主键（导入时会被重置为 0）" },
          "school": { "type": "string" },
          "grade": { "type": "string" },
          "class": { "type": "string" },
          "daily_class": {
            "type": "array",
            "minItems": 7,
            "maxItems": 7,
            "items": { "$ref": "#/components/schemas/DailyClass" }
          }
        }
      },
      "BackupClientConfigRecord": {
        "description": "数据库客户端配置记录",
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "school": { "type": "string" },
          "grade": { "type": "string" },
          "class": { "type": "string" },
          "countdown_target": { "type": "string" },
          "weather_alert_override": { "type": "boolean" },
          "weather_alert_brief": { "type": "boolean" },
          "week_display": { "type": "boolean" },
          "banner_text": { "type": "string" },
          "css_style": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "startup_behavior": {
            "type": "string",
            "enum": ["exit", "stay", "normal"],
            "default": "normal"
          },
          "temperature_colors": {
            "$ref": "#/components/schemas/TemperatureColorsConfig"
          }
        }
      },
      "BackupTimetableRecord": {
        "description": "数据库作息配置记录",
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "school": { "type": "string" },
          "grade": { "type": "string" },
          "timetable": { "$ref": "#/components/schemas/TimetableMap" },
          "divider": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "integer" }
            }
          },
          "start": { "type": "string", "description": "开学日期" }
        }
      },
      "BackupSubjectRecord": {
        "description": "数据库科目配置记录",
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "school": { "type": "string" },
          "grade": { "type": "string" },
          "subject_name": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "科目简称→全称映射"
          }
        }
      },
      "BackupDataVersionRecord": {
        "description": "数据库版本记录",
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "school": { "type": "string" },
          "grade": { "type": "string" },
          "class": { "type": "string" },
          "version": { "type": "string", "format": "date-time", "description": "版本时间戳" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "BackupAutorunRecord": {
        "description": "数据库自动任务记录",
        "type": "object",
        "properties": {
          "hashid": { "type": "string", "description": "HashID（主键）" },
          "etype": { "type": "integer", "description": "任务类型：0=COMPENSATION, 1=TIMETABLE, 2=SCHEDULE, 3=ALL" },
          "scope": {
            "type": "array",
            "items": { "type": "string" }
          },
          "parameters": {
            "type": "object",
            "additionalProperties": true,
            "description": "规则参数"
          },
          "level": { "type": "integer", "description": "优先级" },
          "status": { "type": "integer", "description": "状态：0=待生效, 1=生效中, 2=已过期" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "BackupCountdownRecord": {
        "description": "数据库倒数日配置记录",
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "配置 ID（主键）" },
          "scope": {
            "type": "array",
            "items": { "type": "string" }
          },
          "schedules": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CountdownScheduleItem" }
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "BackupImportResult": {
        "description": "备份导入结果",
        "type": "object",
        "properties": {
          "imported": {
            "type": "object",
            "additionalProperties": { "type": "integer" },
            "description": "各表导入的条数（schedules, client_configs, timetables, subjects, data_versions, autorun_records, countdown_records）"
          },
          "total": { "type": "integer", "description": "总共导入的条数" }
        }
      },
      "ErrorWithErrorField": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "DetailResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      },
...
      "TemperatureColorsConfig": {
        "type": "object",
        "description": "温度颜色配置",
        "properties": {
          "use_gradient": { "type": "boolean", "default": false, "description": "是否使用渐变色" },
          "stops": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TemperatureStop" },
            "description": "温度-颜色断点列表"
          }
        }
      },
      "TemperatureStop": {
        "type": "object",
        "required": ["temp", "color"],
        "properties": {
          "temp": { "type": "number", "description": "温度阈值" },
          "color": { "type": "string", "description": "对应颜色（HEX）" }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "（已废弃，仅供旧版客户端兼容）用户名：`AstraSchedule` 或 `ElectronClassSchedule`，密码：配置文件 `secret.token`"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，通过 POST /web/auth/login 获取。写入接口还需在 X-Verify-Password 头中提供密码。"
      },
      "internalAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Secret",
        "description": "内部 API 密钥，用于服务间调用。配置文件 [internal].secret"
      },
      "regToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Reg-Token",
        "description": "注册 JWT 令牌，由 reg-to 签发，有效期 10 分钟。包含完整注册信息。"
      }
    }
  },
  "servers": [
    {
      "url": "https://{hostvar1}.getastra.cn"
    }
  ],
  "security": []
}